Test and validate on input change
